+++ psse/psse-model/src/test/resources/IEEE_14_bus.json
@@ -951,5 +951,227 @@
     "o4" : 0,
     "f4" : 1.0
   } ],
-  "transformers" : [ ]
+  "transformers" : [ {
+    "i" : 4,
+    "j" : 7,
+    "k" : 0,
+    "ckt" : "1 ",
+    "cw" : 1,
+    "cz" : 1,
+    "cm" : 1,
+    "mag1" : 0.0,
+    "mag2" : 0.0,
+    "nmetr" : 2,
+    "name" : "        ",
+    "stat" : 1,
+    "o1" : 1,
+    "f1" : 1.0,
+    "o2" : 0,
+    "f2" : 1.0,
+    "o3" : 0,
+    "f3" : 1.0,
+    "o4" : 0,
+    "f4" : 1.0,
+    "vecgrp" : null,
+    "starModel" : {
+      "r12" : 0.0,
+      "x12" : 0.20912,
+      "sbase12" : 100.0,
+      "r23" : 0.0,
+      "x23" : 0.0,
+      "sbase23" : 0.0,
+      "r31" : 0.0,
+      "x31" : 0.0,
+      "sbase31" : 0.0,
+      "vmstar" : 0.0,
+      "anstar" : 0.0
+    },
+    "winding1" : {
+      "windv" : 0.978,
+      "nomv" : 0.0,
+      "ang" : 0.0,
+      "rata" : 0.0,
+      "ratb" : 0.0,
+      "ratc" : 0.0,
+      "cod" : 0,
+      "cont" : 0,
+      "rma" : 1.5,
+      "rmi" : 0.51,
+      "vma" : 1.5,
+      "vmi" : 0.51,
+      "ntp" : 159,
+      "tab" : 0,
+      "cr" : 0.0,
+      "cx" : 0.0,
+      "cnxa" : 0.0
+    },
+    "winding2" : {
+      "windv" : 1.0,
+      "nomv" : 0.0,
+      "ang" : 0.0,
+      "rata" : 0.0,
+      "ratb" : 0.0,
+      "ratc" : 0.0,
+      "cod" : 0,
+      "cont" : 0,
+      "rma" : 0.0,
+      "rmi" : 0.0,
+      "vma" : 0.0,
+      "vmi" : 0.0,
+      "ntp" : 0,
+      "tab" : 0,
+      "cr" : 0.0,
+      "cx" : 0.0,
+      "cnxa" : 0.0
+    },
+    "winding3" : null
+  }, {
+    "i" : 4,
+    "j" : 9,
+    "k" : 0,
+    "ckt" : "1 ",
+    "cw" : 1,
+    "cz" : 1,
+    "cm" : 1,
+    "mag1" : 0.0,
+    "mag2" : 0.0,
+    "nmetr" : 2,
+    "name" : "        ",
+    "stat" : 1,
+    "o1" : 1,
+    "f1" : 1.0,
+    "o2" : 0,
+    "f2" : 1.0,
+    "o3" : 0,
+    "f3" : 1.0,
+    "o4" : 0,
+    "f4" : 1.0,
+    "vecgrp" : null,
+    "starModel" : {
+      "r12" : 0.0,
+      "x12" : 0.55618,
+      "sbase12" : 100.0,
+      "r23" : 0.0,
+      "x23" : 0.0,
+      "sbase23" : 0.0,
+      "r31" : 0.0,
+      "x31" : 0.0,
+      "sbase31" : 0.0,
+      "vmstar" : 0.0,
+      "anstar" : 0.0
+    },
+    "winding1" : {
+      "windv" : 0.969,
+      "nomv" : 0.0,
+      "ang" : 0.0,
+      "rata" : 0.0,
+      "ratb" : 0.0,
+      "ratc" : 0.0,
+      "cod" : 0,
+      "cont" : 0,
+      "rma" : 1.5,
+      "rmi" : 0.51,
+      "vma" : 1.5,
+      "vmi" : 0.51,
+      "ntp" : 159,
+      "tab" : 0,
+      "cr" : 0.0,
+      "cx" : 0.0,
+      "cnxa" : 0.0
+    },
+    "winding2" : {
+      "windv" : 1.0,
+      "nomv" : 0.0,
+      "ang" : 0.0,
+      "rata" : 0.0,
+      "ratb" : 0.0,
+      "ratc" : 0.0,
+      "cod" : 0,
+      "cont" : 0,
+      "rma" : 0.0,
+      "rmi" : 0.0,
+      "vma" : 0.0,
+      "vmi" : 0.0,
+      "ntp" : 0,
+      "tab" : 0,
+      "cr" : 0.0,
+      "cx" : 0.0,
+      "cnxa" : 0.0
+    },
+    "winding3" : null
+  }, {
+    "i" : 5,
+    "j" : 6,
+    "k" : 0,
+    "ckt" : "1 ",
+    "cw" : 1,
+    "cz" : 1,
+    "cm" : 1,
+    "mag1" : 0.0,
+    "mag2" : 0.0,
+    "nmetr" : 2,
+    "name" : "        ",
+    "stat" : 1,
+    "o1" : 1,
+    "f1" : 1.0,
+    "o2" : 0,
+    "f2" : 1.0,
+    "o3" : 0,
+    "f3" : 1.0,
+    "o4" : 0,
+    "f4" : 1.0,
+    "vecgrp" : null,
+    "starModel" : {
+      "r12" : 0.0,
+      "x12" : 0.25202,
+      "sbase12" : 100.0,
+      "r23" : 0.0,
+      "x23" : 0.0,
+      "sbase23" : 0.0,
+      "r31" : 0.0,
+      "x31" : 0.0,
+      "sbase31" : 0.0,
+      "vmstar" : 0.0,
+      "anstar" : 0.0
+    },
+    "winding1" : {
+      "windv" : 0.932,
+      "nomv" : 0.0,
+      "ang" : 0.0,
+      "rata" : 0.0,
+      "ratb" : 0.0,
+      "ratc" : 0.0,
+      "cod" : 0,
+      "cont" : 0,
+      "rma" : 1.5,
+      "rmi" : 0.51,
+      "vma" : 1.5,
+      "vmi" : 0.51,
+      "ntp" : 159,
+      "tab" : 0,
+      "cr" : 0.0,
+      "cx" : 0.0,
+      "cnxa" : 0.0
+    },
+    "winding2" : {
+      "windv" : 1.0,
+      "nomv" : 0.0,
+      "ang" : 0.0,
+      "rata" : 0.0,
+      "ratb" : 0.0,
+      "ratc" : 0.0,
+      "cod" : 0,
+      "cont" : 0,
+      "rma" : 0.0,
+      "rmi" : 0.0,
+      "vma" : 0.0,
+      "vmi" : 0.0,
+      "ntp" : 0,
+      "tab" : 0,
+      "cr" : 0.0,
+      "cx" : 0.0,
+      "cnxa" : 0.0
+    },
+    "winding3" : null
+  } ]
 }
